post and get request

+++ src/main/java/org/example/bookmyshow/dtos/BookTicketResponseDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class BookTicketResponseDto {
     private Long bookingId;
     private double amount;
-    private List<Integer> seatNumbers;
+    private List<Long> seatNumbers;
     private String theatreName;
 }
